backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
asyncmy==0.2.9
aiosqlite==0.19.0
greenlet==3.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI & Agents
phidata==2.4.30
google-generativeai==0.3.2

# HTTP & WebSocket
aiohttp==3.9.1
websockets==12.0
httpx==0.26.0

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
